feature(validation): renamed sanity validation files
