+++ packages/groq-builder/src/validation/index.ts
@@ -1,14 +1,7 @@
 import { primitiveValidators } from "./primitives";
-import { createGroqBuilder, RootConfig, GroqBuilderOptions } from "../index";
-import { sanityValidators } from "./sanity";
+import { sanityValidators } from "./sanity-content-blocks";
 
 export const validate = {
   ...primitiveValidators,
   ...sanityValidators,
 };
-
-export function createGroqBuilderWithValidation<TRootConfig extends RootConfig>(
-  options?: GroqBuilderOptions
-) {
-  return Object.assign(createGroqBuilder<TRootConfig>(options), validate);
-}
