+++ packages/groq-builder/src/validation/sanity-content-blocks.ts
